+++ legacy-registries-v0/src/test/java/io/github/minecraftcursedlegacy/test/RegistryTest.java
@@ -2,14 +2,12 @@ package io.github.minecraftcursedlegacy.test;
 
 import io.github.minecraftcursedlegacy.api.registry.Id;
 import io.github.minecraftcursedlegacy.api.registry.Registries;
-import io.github.minecraftcursedlegacy.api.registry.TileEntities;
 import io.github.minecraftcursedlegacy.api.registry.TileItems;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.item.ItemInstance;
 import net.minecraft.item.ItemType;
 import net.minecraft.recipe.SmeltingRecipeRegistry;
 import net.minecraft.tile.Tile;
-import net.minecraft.tile.entity.TileEntity;
 
 public class RegistryTest implements ModInitializer {
 	@Override
@@ -22,20 +20,10 @@ public class RegistryTest implements ModInitializer {
 		tileItem = TileItems.registerTileItem(new Id("modid:tile"), tile);
 
 
-		tileWithEntity = Registries.TILE.register(new Id("modid:tile_with_entity"),
-				i -> new BasicTileWithEntity(i).setName("exampleBlockWithEntity"));
-		tileWithEntityItem = TileItems.registerTileItem(new Id("modid:tile_with_entity"), tileWithEntity);
-		tileEntityClass = BasicTileWithEntity.BasicTileEntity.class;
-		TileEntities.registerTileEntity(tileEntityClass, new Id("modid:tile_entity"));
-
 		SmeltingRecipeRegistry.getInstance().addSmeltingRecipe(item.id, new ItemInstance(tile));
-		SmeltingRecipeRegistry.getInstance().addSmeltingRecipe(Tile.DIRT.id, new ItemInstance(tileWithEntity));
 	}
 
 	public static ItemType item;
 	public static Tile tile;
 	public static ItemType tileItem;
-	public static Tile tileWithEntity;
-	public static ItemType tileWithEntityItem;
-	public static Class<? extends TileEntity> tileEntityClass;
 }
