+++ legacy-tile-entities-v1/build.gradle
@@ -1,0 +1,7 @@
+archivesBaseName = "legacy-tile-entities-v1"
+version = getSubprojectVersion(project, "0.6.3")
+
+dependencies {
+	compile project(path: ':legacy-api-base')
+	compile project(path: ':legacy-registries-v0')
+}
