+++ legacy-tile-entities-v1/src/main/resources/fabric.mod.json
@@ -1,0 +1,34 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-tile-entities",
+  "version": "${version}",
+
+  "name": "Minecraft Cursed Legacy Tile Entities",
+  "description": "A set of useful hooks and features to be used by modders.",
+  "authors": [
+    "The Cursed Legacy Team"
+  ],
+  "contact": {
+    "issues": "https://github.com/minecraft-cursed-legacy/Cursed-Legacy-API/issues",
+    "sources": "https://github.com/minecraft-cursed-legacy/Cursed-Legacy-API"
+  },
+
+  "license": "MIT",
+  "icon": "assets/api/icon.png",
+  
+  "entrypoints": {
+    "init": []
+  },
+  
+  "mixins": [
+    "legacy-tile-entities.accessors.json",
+    "legacy-tile-entities.mixins.json"
+  ],
+
+  "depends": {
+    "fabricloader": "*",
+    "legacy-api-base": "*",
+    "legacy-api-registries-v0": "*",
+    "minecraft": "1.0.0-beta.7.3"
+  }
+}
