+++ legacy-tile-entities-v1/src/main/resources/legacy-registries.accessors.json
@@ -9,6 +9,7 @@
     "AccessorRecipeRegistry",
     "AccessorShapedRecipe",
     "AccessorShapelessRecipe",
+    "AccessorTileEntity",
     "AccessorTileItem"
   ],
   "injectors": {
