+++ legacy-tile-entities-v1/src/main/resources/legacy-registries.mixins.json
@@ -1,0 +1,22 @@
+{
+  "required": true,
+  "package": "io.github.minecraftcursedlegacy.mixin.registry",
+  "compatibilityLevel": "JAVA_8",
+  "mixins": [
+    "MixinItemType",
+    "MixinMcRegionLevelStorage",
+    "MixinOldLevelStorage",
+    "MixinPlaceableTileItem",
+    "MixinShapedRecipe",
+    "MixinSmeltingRecipeRegistry",
+    "MixinTile",
+    "MixinTileItem"
+  ],
+  "client": [
+    "MixinHandItemRenderer",
+    "MixinItemRenderer"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
