+++ legacy-tile-entities-v1/src/test/resources/fabric.mod.json
@@ -1,0 +1,32 @@
+{
+  "schemaVersion": 1,
+  "id": "legacy-registries-test",
+  "version": "${version}",
+
+  "name": "Registries Test Mod",
+  "description": "Tests for the registry module of Cursed Legacy API.",
+  "authors": [
+    "The Cursed Legacy Team"
+  ],
+  "contact": {
+    "issues": "https://github.com/minecraft-cursed-legacy/Cursed-Legacy-API/issues",
+    "sources": "https://github.com/minecraft-cursed-legacy/Cursed-Legacy-API"
+  },
+
+  "license": "MIT",
+  "icon": "assets/modid/icon.png",
+
+  "environment": "*",
+  "entrypoints": {
+    "init": [
+      "io.github.minecraftcursedlegacy.test.TileEntityTest"
+    ]
+  },
+
+  "depends": {
+    "fabricloader": "*",
+    "legacy-api-base": "*",
+    "legacy-registries": "*",
+    "minecraft": "1.0.0-beta.7.3"
+  }
+}
