+++ settings.gradle
@@ -22,5 +22,6 @@ include 'legacy-networking-v0'
 include 'legacy-recipes-v0'
 include 'legacy-registries-v0'
 include 'legacy-terrain-v1'
+include 'legacy-tile-entities-v1'
 include 'legacy-translations-v0'
 include 'legacy-worldtypes-v1'
